+++ spec/app_templates/testapp/app/controllers/home_controller.rb
@@ -1,9 +1,5 @@
 class HomeController < ApplicationController
   def show
-    if Rails::VERSION::MAJOR >= 5
-      render html: "", layout: "application"
-    else
-      render text: "", layout: "application"
-    end
+    render html: "", layout: "application"
   end
 end